# openapi.yaml
openapi: 3.1.0
info:
  title: saas-scaffolding
  version: 1.0.0
  description: Multi-tenant SaaS scaffolding API with authentication, organization management, microservice tokens, and webhook integrations.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie set after successful sign-in.
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal server error
    ErrorResponseWithError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Unauthorized
...
    UpdateOrganizationRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: New Org Name
    UpdatedOrganization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
    UpdateOrganizationSettingsRequest:
      type: object
      required:
        - autoAcceptDomainMembers
      properties:
        autoAcceptDomainMembers:
          type: boolean
          description: Enable or disable automatic membership approval for users with a matching email domain.
    UpdateOrganizationSettingsResponse:
      type: object
      required:
        - settings
      properties:
        settings:
          type: object
          properties:
            enabled:
              type: boolean
            domain:
              type:
                - string
                - "null"
    LogoUploadResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Logo uploaded successfully
        logoUrl:
          type: string
          format: uri
...
    MicroserviceTokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - tokenType
      properties:
        accessToken:
          type: string
          description: JWT access token for microservice calls.
        refreshToken:
          type: string
          description: Opaque refresh token.
        expiresIn:
          type: integer
          description: Token lifetime in seconds.
          example: 3600
        tokenType:
          type: string
          example: Bearer
...
    RevokeTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Required unless revokeAll is true.
        revokeAll:
          type: boolean
          description: When true, revoke all refresh tokens for the authenticated user.
...
  responses:
    BadRequest:
      description: Bad request - validation error or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or session invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/organizations/{organizationId}:
    patch:
      operationId: updateOrganization
      summary: Rename an organization
      description: |
        Updates the organization name. Requires owner or admin role.
        Rate limit tier: apiRateLimit.
      tags:
        - organizations
      security:
        - sessionCookie: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedOrganization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Organizations - Settings
  # ──────────────────────────────────────────────
  /api/organizations/{organizationId}/settings:
    patch:
      operationId: updateOrganizationSettings
      summary: Update organization settings
      description: |
        Toggles auto-accept domain membership. When enabled, users whose email domain matches the organization domain are automatically approved.
        Personal email domains (gmail.com, etc.) are rejected. Requires owner or admin role.
        Rate limit tier: apiRateLimit.
      tags:
        - organizations
      security:
        - sessionCookie: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationSettingsRequest"
      responses:
        "200":
          description: Settings updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateOrganizationSettingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Organizations - Logo Upload
  # ──────────────────────────────────────────────
  /api/organizations/{organizationId}/logo:
    post:
      operationId: uploadOrganizationLogo
      summary: Upload organization logo
      description: |
        Uploads a new logo image for the organization. Replaces the previous logo if one exists.
        Requires owner or admin role.
        Rate limit tier: uploadRateLimit.
      tags:
        - organizations
      security:
        - sessionCookie: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG, JPG, etc.).
      responses:
        "200":
          description: Logo uploaded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # Tenants - Audit Logs
  # ──────────────────────────────────────────────
...
  /api/microservices/token:
    post:
      operationId: generateMicroserviceToken
      summary: Generate microservice access and refresh tokens
      description: |
        Issues a JWT access token (1 hour) and an opaque refresh token scoped to the user's current tenant organization.
        The tenant is resolved from the request context (subdomain or header).
        Rate limit tier: apiRateLimit.
      tags:
        - microservices
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Token pair generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MicroserviceTokenResponse"
        "400":
          description: No organization context found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"

  # ──────────────────────────────────────────────
  # Microservices - Refresh Token
  # ──────────────────────────────────────────────
...
  /api/microservices/revoke:
    post:
      operationId: revokeMicroserviceToken
      summary: Revoke refresh token(s)
      description: |
        Revokes a specific refresh token or all refresh tokens for the authenticated user (when `revokeAll` is true).
        Rate limit tier: apiRateLimit.
      tags:
        - microservices
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeTokenRequest"
      responses:
        "200":
          description: Token(s) revoked.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token revoked successfully
        "400":
          description: Missing refresh token (when revokeAll is not set).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseWithError"

  # ──────────────────────────────────────────────
  # Users - Avatar
  # ──────────────────────────────────────────────
...